fix: address review feedback on LLM client retry, markdown lint, and tokio features
- docs/llm-client.md: fix MD022/MD031/MD040 by adding blank lines after
  level-3 headings and language tags to fenced code blocks
- mimir-core/Cargo.toml: narrow tokio features to rt-multi-thread+time in
  dependencies; add tokio with macros+rt-multi-thread to dev-dependencies
- mimir-core/src/llm/client.rs:
  * Clone now clones the existing reqwest::Client to preserve pool/config
  * send_request returns Result<Response, LlmError> and converts non-success
    HTTP statuses into LlmError::Api so transient codes reach retry logic
  * retry_with_backoff/is_transient updated to retry both Network errors and
    transient Api errors (429/502/503/504)
  * Remove unnecessary as u32 cast by aligning LlmConfig.max_tokens to u32
- mimir-core/src/config.rs: change LlmConfig.max_tokens from u16 to u32

+++ mimir-core/src/config.rs
@@ -21,7 +21,7 @@ pub struct LlmConfig {
     pub endpoint: String,
     pub api_key: String,
     pub model: String,
-    pub max_tokens: u16,
+    pub max_tokens: u32,
     pub temperature: f32,
 }
 
@@ -195,7 +195,7 @@ impl Config {
             self.llm.model = val;
         }
         if let Ok(val) = std::env::var("MIMIR_LLM_MAX_TOKENS")
-            && let Ok(parsed) = val.parse::<u16>()
+            && let Ok(parsed) = val.parse::<u32>()
         {
             self.llm.max_tokens = parsed;
         }
